requirements.txt
Django==1.7
djangorestframework==2.3.13
mock==1.0.1
